requirements_dev.txt
pip>=20.3.1,!=20.3.2
wheel==0.36.2
twine==3.3.0
bump2version==1.0.1
flake8==3.7.8
tox==3.14.0
coverage==5.3.1
pytest==4.6.5
pytest-runner==5.1
Click==7.1.2
antlr4-python3-runtime==4.8